+++ packages/utils/functions/fetch/index.ts
@@ -1,0 +1,2 @@
+export * from './baseAppFetch'
+export type * from './baseAppFetch/types'